Safari-609.1.11.1/LayoutTests/fast/css/outline-offset-parsing-expected.txt

This tests checks parsing of the 'outline-offset' property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("outline-offset: 1px", "outline-offset") is "1px"
PASS test("outline-offset: 1mm", "outline-offset") is "1mm"
PASS test("outline-offset: 1%", "outline-offset") is ""
PASS successfullyParsed is true

TEST COMPLETE
